Added profile picture and photo filtering

+++ src/main/java/miniTwitter/demo/controllers/PhotoController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
@@ -40,9 +41,14 @@ public class PhotoController {
     private PhotoRepository photoRepository;
     
 	
-	   @GetMapping("/upload")
+    	@GetMapping("/upload")
 	    public String uploadForm(){
 	        return "upload";
+	    }
+	   
+	   	@GetMapping("/uploadProfile")
+	    public String uploadProfileForm(){
+	        return "uploadProfile";
 	    }
 
 	    @PostMapping("/upload")
@@ -58,41 +64,88 @@ public class PhotoController {
 
 	            model.addAttribute("message",
 	                    "You successfully uploaded '" + file.getOriginalFilename() + "'");
-	            model.addAttribute("imageurl", uploadResult.get("url"));
+	            
 	            String filename = uploadResult.get("public_id").toString() + "." + uploadResult.get("format").toString();
 	            
-	            
-	            User user = userRepository.findByEmail(p.getName());
-	            
-	            Photo photo = new Photo();
-	            photo.setImage(cloudc.createUrl(filename, 100, 150, "fit", "sepia"));
-	            photo.setFileName(filename);
-	            photo.setCreatedAt(new Date());
-	            photo.setUser(user);
-	            photoRepository.save(photo);
 	           
-	            
+	            model.addAttribute("imageurl", uploadResult.get("url"));
 	            model.addAttribute("imagename", filename);
 	            
 	        } catch (IOException e){
 	            e.printStackTrace();
 	            model.addAttribute("message", "Sorry I can't upload that!");
 	        }
-	        return "upload";
+	        return "filter";
+	    }
+	    
+	    @PostMapping("/uploadProfile")
+	    public String profileUpload(@RequestParam("file") MultipartFile file, RedirectAttributes redirectAttributes,Principal p, Model model){
+
+	        if (file.isEmpty()){
+	            redirectAttributes.addFlashAttribute("message","Please select a file to upload");
+	            return "redirect:uploadStatus";
+	        }
+
+	        try {
+	            Map uploadResult =  cloudc.upload(file.getBytes(), ObjectUtils.asMap("resourcetype", "auto"));
+
+	            model.addAttribute("message",
+	                    "You successfully uploaded '" + file.getOriginalFilename() + "'");
+	            
+	            String filename = uploadResult.get("public_id").toString() + "." + uploadResult.get("format").toString();
+	            
+	            User user = userRepository.findByEmail(p.getName());
+	            
+	            Photo photo = new Photo();
+	            photo.setImage(cloudc.createautoUrl(filename, 150, 150));
+	            photo.setFileName(filename);
+	            photo.setCreatedAt(new Date());
+	            photo.setUser(user);
+	            photoRepository.save(photo);
+	            
+	            user.setProfilePicture(photo);
+	            userRepository.save(user);
+	            
+	            model.addAttribute("imageurl", cloudc.createautoUrl(filename, 150, 150));
+	            model.addAttribute("user", user);
+	            model.addAttribute("imagename", filename);
+	            
+	        } catch (IOException e){
+	            e.printStackTrace();
+	            model.addAttribute("message", "Sorry I can't upload that!");
+	        }
+	        return "tweet";
 	    }
 	    
 	    @RequestMapping("/filter")
-	    public String filter(String imagename, int width, int height, String action,String filter, Model model){
+	    public String filter(String imagename, int width, int height, String action,String filter,Principal principal, Model model){
 	    	
 	    	if(width==0 ){
-	    		width=150;
+	    		width=250;
 	    	}
 	    	if(height==0){
-	    		height=150;
+	    		height=250;
 	    	}
-	    	 model.addAttribute("sizedimageurl", cloudc.createUrl(imagename, width, height, action, filter));
+	    	
+	    	model.addAttribute("sizedimageurl", cloudc.createUrl(imagename, width, height, action, filter));
+	    	model.addAttribute("imagename", imagename);
 	    	 
-	    	 return "upload";
+	    	return "filter";
+	    }
+	    
+	    
+	    @RequestMapping(path="/save" , method=RequestMethod.POST)
+	    public String save(String sizedimageurl, String imagename, Principal principal, Model model){
+	    	
+	    	Photo photo = new Photo();
+            photo.setImage(sizedimageurl);
+            photo.setFileName(imagename);
+            photo.setCreatedAt(new Date());
+            User user = userRepository.findByEmail(principal.getName());
+            photo.setUser(user);
+            photoRepository.save(photo);
+	    	 
+	    	return "redirect:/newsfeed";
 	    }
 	    
 	    @RequestMapping(path="/getphotos")
@@ -105,7 +158,7 @@ public class PhotoController {
 	    	model.addAttribute("user", user);
 	    	model.addAttribute("photos", pictures);
 	  
-	    	return "gallery";
+	    	return "newsfeed";
 	    }
 	    
 	    @RequestMapping(path="/delete/{photoId}")
@@ -113,13 +166,13 @@ public class PhotoController {
 	    	
 	    	User user = userRepository.findByEmail(p.getName());
 	    	Photo photo = photoRepository.findOne(photoId);
-	    	photoRepository.delete(photo);
+	    	if(photo.getUser() == user){
+		    	
+	    		photoRepository.delete(photo);
+		    	
+		    } 
 	    	
-	    	List<Photo> pictures = photoRepository.findByUser_Id(user.getId());
-	    	model.addAttribute("user", user);
-	    	model.addAttribute("photos", pictures);
-	    	
-	    	return "gallery";
+	    	return "redirect:/newsfeed";
 	    }
 
 }
